Api/requirements.txt
fastapi
uvicorn
pillow
numpy
tensorflow
scikit-learn   # Kullandığın modele göre değişebilir: torch, tensorflow vs.
python-multipart
opencv-python